chore: move metadata types to tractusx package
