+++ pkg/tractusx/metadata_test.go
@@ -17,7 +17,7 @@
  * SPDX-License-Identifier: Apache-2.0
  ******************************************************************************/
 
-package product
+package tractusx
 
 import (
 	"os"
