+++ pkg/tractusx/test/metadata_test_template.yaml
@@ -1,0 +1,29 @@
+###############################################################################
+# Copyright (c) 2023 Contributors to the Eclipse Foundation
+#
+# See the NOTICE file(s) distributed with this work for additional
+# information regarding copyright ownership.
+#
+# This program and the accompanying materials are made available under the
+# terms of the Apache License, Version 2.0 which is available at
+# https://www.apache.org/licenses/LICENSE-2.0.
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+# WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+# License for the specific language governing permissions and limitations
+# under the License.
+#
+# SPDX-License-Identifier: Apache-2.0
+##############################################################################
+---
+
+product: "sig-infra"
+leadingRepository: "https://github.com/eclipse-tractusx/sig-infra"
+repositories:
+  - name: "sig-infra"
+    usage: "Contains issue templates and workflows"
+    url: "https://github.com/eclipse-tractusx/sig-infra"
+  - name: "charts"
+    usage: "Central Helm repository containing all released charts of Tractus-X"
+    url: "https://github.com/eclipse-tractusx/charts"
